+++ programs/dynamic-bonding-curve/src/instructions/creator/mod.rs
@@ -6,7 +6,3 @@ pub mod ix_withdraw_creator_surplus;
 pub use ix_withdraw_creator_surplus::*;
 pub mod ix_transfer_pool_creator;
 pub use ix_transfer_pool_creator::*;
-pub mod ix_claim_spl_token_update_authority;
-pub use ix_claim_spl_token_update_authority::*;
-pub mod ix_claim_token_2022_update_authority;
-pub use ix_claim_token_2022_update_authority::*;