+++ tests/create_virtual_pool_metadata.tests.ts
@@ -98,6 +98,7 @@ describe("Create virtual pool metadata", () => {
             migrationFeeOption: 0,
             tokenSupply: null,
             creatorTradingFeePercentage: 0,
+            tokenUpdateAuthority: 0,
             padding0: [],
             padding: [],
             curve: curves,